create a service
